Create MDT Version 1
This is an MDT (Mobile Data Tablet) version 1.0. I was inspired to do this project by a game called ER:LC (Emergency Response : Liberty County) in Roblox. I will put a link to the game because of how realistic the game is. I am a really big fan of the game and has been playing it for a couple years. The reason I started coding is because of the creator, mrfergie. I will also put the link for his profile and a link to the Roblox group PRC (Police Roleplay Community) which he owns and the group which made the game with him. Please make sure to check them out.


The Game: https://www.roblox.com/games/2534724415/Emergency-Response-Liberty-County

The Group: https://www.roblox.com/communities/4328109/Police-Roleplay-Community#!/about

The Owner: https://www.roblox.com/users/7205597/profile